backend/saint_analysis/requirements.txt
fastapi
uvicorn
python-dotenv
torch
tqdm
scikit-learn
pandas
numpy
pymilvus
requests

